Simplify: drop shared event render cache, keep call_soon dispatch fix
Benchmarks showed event serialization costs only 9-36us per event, so the
per-(event, locale) render cache added complexity for negligible gain with
the typical handful of connected clients.
