+++ tests/controllers/webserver/test_event_dispatch.py
@@ -1,0 +1,91 @@
+"""Tests for event dispatch and fan-out towards websocket clients."""
+
+from __future__ import annotations
+
+import asyncio
+from types import SimpleNamespace
+from typing import TYPE_CHECKING
+
+import pytest
+from aiohttp import web
+from aiohttp.test_utils import make_mocked_request
+from music_assistant_models.auth import User, UserRole
+from music_assistant_models.enums import EventType
+
+from music_assistant.controllers.webserver.controller import WebserverController
+from music_assistant.controllers.webserver.websocket_client import WebsocketClientHandler
+
+if TYPE_CHECKING:
+    from music_assistant.mass import MusicAssistant
+
+
+async def drain_event_callbacks() -> None:
+    """Yield to the event loop so event subscriber callbacks (call_soon) run."""
+    await asyncio.sleep(0)
+
+
+@pytest.fixture
+def webserver(mass_minimal: MusicAssistant) -> WebserverController:
+    """Return a WebserverController with stubbed serialization dependencies."""
+    # stub the controllers referenced by the serialization resolvers
+    # (mass_minimal does not set up metadata/translations/tasks)
+    mass_minimal.metadata = SimpleNamespace(  # type: ignore[assignment]
+        compute_image_id=lambda provider, path: f"{provider}--{path}"
+    )
+    mass_minimal.translations = SimpleNamespace(  # type: ignore[assignment]
+        get_translation=lambda _key, **_kwargs: None
+    )
+    webserver = WebserverController(mass_minimal)
+    mass_minimal.webserver = webserver
+    return webserver
+
+
+def create_ws_client(webserver: WebserverController, username: str) -> WebsocketClientHandler:
+    """Create an authenticated + event-subscribed websocket client handler (no real socket)."""
+    request = make_mocked_request("GET", "/ws", app=web.Application())
+    client = WebsocketClientHandler(webserver, request)
+    client._authenticated_user = User(user_id=username, username=username, role=UserRole.ADMIN)
+    client._subscribe_to_events()
+    return client
+
+
+def get_written_message(client: WebsocketClientHandler) -> str:
+    """Pop the next message queued for the client's writer."""
+    message = client._to_write.get_nowait()
+    assert message is not None
+    return message
+
+
+async def test_event_delivered_to_all_clients(
+    mass_minimal: MusicAssistant,
+    webserver: WebserverController,
+) -> None:
+    """An event signalled on the loop thread reaches every subscribed websocket client."""
+    client1 = create_ws_client(webserver, "user1")
+    client2 = create_ws_client(webserver, "user2")
+
+    mass_minimal.signal_event(EventType.PLAYER_UPDATED, "player1", {"name": "Test Player"})
+    await drain_event_callbacks()
+
+    msg1 = get_written_message(client1)
+    msg2 = get_written_message(client2)
+    assert msg1 == msg2
+    assert "player1" in msg1
+
+
+async def test_tasks_updated_rendered_per_user(
+    mass_minimal: MusicAssistant,
+    webserver: WebserverController,
+) -> None:
+    """TASKS_UPDATED events carry a per-user payload."""
+    mass_minimal.tasks = SimpleNamespace(  # type: ignore[assignment]
+        list_tasks_for_user=lambda user: [{"task_id": f"task-for-{user.username}"}]
+    )
+    client1 = create_ws_client(webserver, "user1")
+    client2 = create_ws_client(webserver, "user2")
+
+    mass_minimal.signal_event(EventType.TASKS_UPDATED)
+    await drain_event_callbacks()
+
+    assert "task-for-user1" in get_written_message(client1)
+    assert "task-for-user2" in get_written_message(client2)
